Extract `QueryDetails` to its own package to avoid a circular dependency between packages
