+++ pkg/frontend/querymiddleware/generic_query_cache.go
@@ -15,6 +15,7 @@ import (
 	"github.com/grafana/dskit/tenant"
 
 	apierror "github.com/grafana/mimir/pkg/api/error"
+	"github.com/grafana/mimir/pkg/frontend/querymiddleware/querydetails"
 	"github.com/grafana/mimir/pkg/streamingpromql/optimize/plan/splitandcache"
 	"github.com/grafana/mimir/pkg/streamingpromql/requestoptions"
 	"github.com/grafana/mimir/pkg/util/spanlogger"
@@ -162,7 +163,7 @@ func (c *genericQueryCache) storeCachedResponse(ctx context.Context, cachedRes *
 }
 
 func (c *genericQueryCache) recordCacheHitQueryDetails(ctx context.Context, hits map[string][]byte) {
-	details := QueryDetailsFromContext(ctx)
+	details := querydetails.QueryDetailsFromContext(ctx)
 	if details == nil {
 		return
 	}
@@ -172,7 +173,7 @@ func (c *genericQueryCache) recordCacheHitQueryDetails(ctx context.Context, hits
 }
 
 func (c *genericQueryCache) recordCacheStoreQueryDetails(ctx context.Context, value []byte) {
-	details := QueryDetailsFromContext(ctx)
+	details := querydetails.QueryDetailsFromContext(ctx)
 	if details == nil {
 		return
 	}
